refactor(database): Split galaxy coord and star class helpers
Move coordinate normalization and star class encoding out of the parser so build and query code can share the same helpers. Centralize Hilbert curve initialization to avoid rebuilding curve state in each consumer.
